根据玩童的isbn从豆瓣获取信息  v1.0

+++ src/main/java/com/booyue/migu/dao/MyBatisBaseDao.java
@@ -13,7 +13,7 @@ import java.util.List;
  */
 public interface MyBatisBaseDao<Model, PK extends Serializable, E> {
 
-    long countByExample(E example);
+    int countByExample(E example);
 
     int deleteByExample(E example);
